+++ packages/browser-injected-sdk/src/solana/signAndSendTransaction.test.ts
@@ -1,4 +1,4 @@
-import type { Transaction } from "@solana/web3.js";
+import type { VersionedTransaction } from "@solana/web3.js";
 import type { SolanaStrategy } from "./strategies/types";
 import { getProvider } from "./getProvider";
 import { signAndSendTransaction } from "./signAndSendTransaction";
@@ -8,7 +8,9 @@ jest.mock("./getProvider", () => ({
   getProvider: jest.fn(),
 }));
 
-const mockTransaction = {} as Transaction;
+const mockTransaction = {
+  serialize: jest.fn().mockReturnValue(new Uint8Array([])),
+} as any as VersionedTransaction;
 
 describe("signAndSendTransaction", () => {
   let mockProvider: Partial<PhantomSolanaProvider>;