+++ packages/indexed-db-stamper/CHANGELOG.md
@@ -1,13 +1,20 @@
 # @phantom/indexed-db-stamper
 
-## 0.1.5
+## 1.0.0-beta.0
+
+### Major Changes
+
+- New SDKs chain specific interface, prepare for initial beta release
 
 ### Patch Changes
 
 - ef68eaf: Key rotation
 - Updated dependencies [ef68eaf]
-  - @phantom/embedded-provider-core@0.1.9
-  - @phantom/sdk-types@0.1.5
+- Updated dependencies
+  - @phantom/embedded-provider-core@1.0.0-beta.0
+  - @phantom/sdk-types@1.0.0-beta.0
+  - @phantom/base64url@1.0.0-beta.0
+  - @phantom/crypto@1.0.0-beta.0
 
 ## 0.1.4