+++ packages/sdk-types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @phantom/sdk-types
 
+## 1.0.0-beta.0
+
+### Major Changes
+
+- New SDKs chain specific interface, prepare for initial beta release
+
 ## 0.1.5
 
 ### Patch Changes